requirements.txt
flask==3.0.3
gunicorn==22.0.0
python-docx==1.1.2
